os/board/rtl8730e: fix 8.4d porting related issues and warnings
*increase max queue number for BLE indication queue
*fix BLE related warnings
*add .gitignore file to prevent missing
*move header files and shorten CFLAGS string
*add missing wifi security types
*fix mixed log issue
*move km4 version to ap layer
*Fix OCF Issue
*change to use zmalloc instead of malloc in ipc api

+++ os/board/rtl8730e/src/component/bluetooth/driver/platform/amebad2/hci/hci_uart.c
@@ -334,7 +334,7 @@ static uint8_t amebad2_uart_open(void)
 	/* Disable and Enable UART Interrupt */
 	InterruptDis(HCI_UART_IRQ);
 	InterruptUnRegister(HCI_UART_IRQ);
-	InterruptRegister((IRQ_FUN)amebad2_uart_irq, HCI_UART_IRQ, NULL, HCI_UART_IRQ_PRIO);
+	InterruptRegister((IRQ_FUN)amebad2_uart_irq, HCI_UART_IRQ, (u32)NULL, HCI_UART_IRQ_PRIO);
 	InterruptEn(HCI_UART_IRQ, HCI_UART_IRQ_PRIO);
 
 	UART_INTConfig(HCI_UART_DEV, RUART_BIT_ETBEI, DISABLE);
